did spotless apply

+++ src/main/java/frc/robot/Constants.java
@@ -60,7 +60,7 @@ public final class Constants {
     public static final double WHEEL_DIAMETER_METERS = 0.15;
     public static final double ENCODER_DISTANCE_PER_PULSE =
         // Assumes the encoders are directly mounted on the wheel shafts
-        (WHEEL_DIAMETER_METERS * Math.PI) / (double) ENCODER_CPR;
+        (WHEEL_DIAMETER_METERS * Math.PI) / ENCODER_CPR;
 
     // These are example values only - DO NOT USE THESE FOR YOUR OWN ROBOT!
     // These characterization values MUST be determined either experimentally or theoretically
